requirements.txt
Flask==0.12.2
gunicorn==19.7.1
six==1.10.0
google-cloud-vision==0.27.0
google-cloud-storage==1.4.0
google-cloud-datastore==1.3.0
